// src/models/TripPlanner/models/TripRequestResponseJourneyLegStop.ts
export interface TripRequestResponseJourneyLegStop {
    /**
     * A timestamp in `YYYY-MM-DDTHH:MM:SSZ` format that indicates the estimated arrival time. If real-time
     * information is available then this timestamp is the real-time estimate, otherwise it is the same as the
     * value in `arrivalTimePlanned`.
     *
     * @type {string}
     * @memberof TripRequestResponseJourneyLegStop
     */
    arrivalTimeEstimated?: string;
    /**
     * A timestamp in `YYYY-MM-DDTHH:MM:SSZ` format that indicates the planned arrival time. This is
     * the original scheduled time.
     *
     * @type {string}
     * @memberof TripRequestResponseJourneyLegStop
     */
    arrivalTimePlanned?: string;
    /**
     * Contains exactly two values: the first value is the latitude, the second value is the longitude.
     *
     * @type {Array<number>}
     * @memberof TripRequestResponseJourneyLegStop
     */
    coord?: Array<number>;
    /**
     * A timestamp in `YYYY-MM-DDTHH:MM:SSZ` format that indicates the estimated departure time. If real-time
     * information is available then this timestamp is the real-time estimate, otherwise it is the same as the
     * value in `departureTimePlanned`.
     *
     * @type {string}
     * @memberof TripRequestResponseJourneyLegStop
     */
    departureTimeEstimated?: string;
    /**
     * A timestamp in `YYYY-MM-DDTHH:MM:SSZ` format that indicates the planned departure time. This is
     * the original scheduled time.
     *
     * @type {string}
     * @memberof TripRequestResponseJourneyLegStop
     */
    departureTimePlanned?: string;
    /**
     * This is the short version of the location name, which does not include the suburb or other information.
     *
     * @type {string}
     * @memberof TripRequestResponseJourneyLegStop
     */
    disassembledName?: string;
    /**
     * This is a unique ID for the returned location. Certain types of ID can be used for subsequent searches performed with `stop_finder`,
     * or can be used as the origin or destination in an `trip` request. The format of a location ID differs greatly, depending
     * on the type of location it is.
     *
     * @type {string}
     * @memberof TripRequestResponseJourneyLegStop
     */
    id?: string;
    /**
     * This is the long version of the location name, which may include the suburb or other information.
     *
     * @type {string}
     * @memberof TripRequestResponseJourneyLegStop
     */
    name?: string;
    /**
     *
     * @type {ParentLocation}
     * @memberof TripRequestResponseJourneyLegStop
     */
    parent?: ParentLocation;
    /**
     *
     * @type {TripRequestResponseJourneyLegStopProperties}
     * @memberof TripRequestResponseJourneyLegStop
     */
    properties?: TripRequestResponseJourneyLegStopProperties;
    /**
     * This is the type of location being returned. It will typically represent a specific stop or platform.
     *
     * @type {string}
     * @memberof TripRequestResponseJourneyLegStop
     */
    type?: TripRequestResponseJourneyLegStopTypeEnum;
}

/**
 * @export
 */
export const TripRequestResponseJourneyLegStopTypeEnum = {
    Poi: "poi",
    Singlehouse: "singlehouse",
    Stop: "stop",
    Platform: "platform",
    Street: "street",
    Locality: "locality",
    Suburb: "suburb",
    Unknown: "unknown",
} as const;
export type TripRequestResponseJourneyLegStopTypeEnum =
    (typeof TripRequestResponseJourneyLegStopTypeEnum)[keyof typeof TripRequestResponseJourneyLegStopTypeEnum];